requirements.txt
python-telegram-bot[job-queue]==21.10
python-dotenv==1.0.1
openai>=1.68.0
dateparser==1.2.0
